Fix DeafultArgProcessor ancd refactoring
