+++ krules_core/tests/subject/test_storaged_subject.py
@@ -114,9 +114,9 @@ def test_set_get_del(subject):
 
     # events
     assert len(_test_events) == 3
-    from krules_core import types
+    from krules_core import event_types
     #   type
-    assert _test_events[0][0] == _test_events[1][0] == _test_events[2][0] == types.SUBJECT_PROPERTY_CHANGED
+    assert _test_events[0][0] == _test_events[1][0] == _test_events[2][0] == event_types.SUBJECT_PROPERTY_CHANGED
     #   subject
     assert _test_events[0][1].name == _test_events[1][1].name == _test_events[2][1].name == subject.name
     #   payload
@@ -207,7 +207,7 @@ def test_set_get_del(subject):
     #   cache not loaded yet
     subject.delete("my-prop", use_cache=False)
     assert len(_test_events) == 1 and \
-        _test_events[0][0] == types.SUBJECT_PROPERTY_DELETED and \
+        _test_events[0][0] == event_types.SUBJECT_PROPERTY_DELETED and \
         _test_events[0][1].name == subject.name and \
         _test_events[0][2][PayloadConst.PROPERTY_NAME] == "my-prop"
     with pytest.raises(AttributeError):
